feat: detect param name from comments and change options
When the option `detectNamedParamFromComments` is `true` (the default), the plugin detects the
parameter names from the `@param` documentation comments. When it is `false`, and
`changeNamedParamTag` is `true` (the default for the latter option), the plugin changes the
parameter names in the documentation tag instead. The options have also been changed from snake-case
to pascalCase for consistently with TypeDoc's options.

BREAKING CHANGE: The `param-name` option has been renamed to `namedParamName` and by default, the
plugin will detect the name of parameters from the documentation comments.